consolidate npe2 stuff
