+++ napari/utils/theme.py
@@ -1,6 +1,5 @@
 # syntax_style for the console must be one of the supported styles from
 # pygments - see here for examples https://help.farbox.com/pygments.html
-import contextlib
 import re
 import warnings
 from ast import literal_eval
@@ -18,6 +17,7 @@ except Exception:
     use_gradients = False
 
 from .._vendor import darkdetect
+from ..plugins._npe2 import install_npe2_themes
 from ..utils.translations import trans
 from .events import EventedModel
 from .events.containers._evented_dict import EventedDict
@@ -320,15 +320,6 @@ _themes: EventedDict[str, Theme] = EventedDict(
     basetype=Theme,
 )
 
-with contextlib.suppress(ImportError):
-    from npe2 import PluginManager
-
-    for theme in PluginManager.instance().iter_themes():
-        # `theme.type` is dark/light and supplies defaults for keys that
-        # are not provided by the plugin
-        d = _themes[theme.type].dict()
-        d.update(theme.colors.dict(exclude_unset=True))
-        _themes[theme.id] = Theme(**d)
-
+install_npe2_themes(_themes, Theme)
 _themes.events.added.connect(rebuild_theme_settings)
 _themes.events.removed.connect(rebuild_theme_settings)
